requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
slackclient==2.0.0
urllib3==1.25.3
websocket-client==0.56.0
yarl==1.1.0
